Organize Bluetooth SDK native sources
